+++ RELEASENOTES.md
@@ -5,6 +5,7 @@
 * Update to Mockito 2
 * Add new `ExoPlaybackException` types for remote exceptions and out-of-memory
   errors.
+* DASH: Parse role and accessibility descriptors into `Format.roleFlags`.
 * HLS:
   * Work around lack of LA_URL attribute in PlayReady key request init data.
   * Prevent unnecessary reloads of initialization segments.